+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to the "project-hierarchy-explorer" extension will be documented in this file.
 
+## [1.0.4]
+
+- improved searchability of the extension
+
+## [1.0.3]
+
+- new configuration -> `outputsTo`, allows you to output your hierarchy to a file, the output console, or both
+- cleaned up and optimized
+
 ## [1.0.2]
 
 - Version bump and sync